bluez: Add chromeos-base/metrics as a dependency and inherit from libchrome
This patch is the initial patch to import the UMA metric API to BlueZ where
chromeos-base/metrics is added as a dependency and libchrome is inherited. The
inheritance of libchrome is for acquiring BASE_VER environment variable in order
to access libmetrics.so.

BUG=chromium:724720
TEST=embed a sample in the start-up of adapter, build BlueZ with lmetrics
linkage and verify the record on chrome://histograms

Change-Id: I31f5d17c5730c2cd20bf9b844c07593b8744f5d5
Reviewed-on: https://chromium-review.googlesource.com/509940
Commit-Ready: Miao-chen Chou <mcchou@chromium.org>
Tested-by: Miao-chen Chou <mcchou@chromium.org>
Reviewed-by: Miao-chen Chou <mcchou@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>