net-print/cups: allow search filter in component mount point.
This allows cupsd/cupstestppd to search filters in non-standard locations (in addition to
/usr/libexec/cups/filter and absolute path). It searches in
/run/imageloader/epson-injket-printer-escpr which holds all mount points of
each component version. cupsdSearchFilterLatest iterates all subdirs in this
folder and compare for the latest version. Then cupsdSearchFilter recursively
search in the latest version folder for the abs path
of filter with 'filter_name'.

For cupsd, it performs search when a filter is not found at standard location.
For cupstestppd, it performs search when a filter is not found at standard location and
skips capitalization test since it fails on squashfs when readdir a string not trailing
with '/'.

BUG=chromium:703896
TEST=run cupstestppd to verify a ppd file requiring escpr filter.

Change-Id: Ia91d1f643680208e61a12e5ab9a1e6c07e85736c
Reviewed-on: https://chromium-review.googlesource.com/518931
Commit-Ready: Xiaochu Liu <xiaochu@chromium.org>
Tested-by: Xiaochu Liu <xiaochu@chromium.org>
Reviewed-by: Xiaochu Liu <xiaochu@chromium.org>