cros-kernel2: use a finer granularity of versions of AutoFDO profiles
Originally all kernels from different branches, patch levels, shared
the same profile, as long as they are based on the same official X.Y
release.  Now make them distinguishable by Chrome OS builds.

For example, there was only one profile for all 4.4 release:
  autofdo-chromeos-kernel-4_4.tar.xz

Now it becomes:
  autofdo-chromeos-kernel-4_4-chromeos-intel-pineview-9460.50.0.afdo.xz

Note that this is a part of the experiemnt (see BUG below).  The naming
convention, along with all design and implmenetation details,  will be
revisited only when the result of this experiemnt is satisfying and it
is decided to roll out for all boards.

BUG=b:32371911
TEST=emerge-quawks chromeos-kernel-4_4

Change-Id: I85e66641dc92586132556afd47b656a82c336180
Reviewed-on: https://chromium-review.googlesource.com/522965
Commit-Ready: Ting-Yuan Huang <laszio@chromium.org>
Tested-by: Ting-Yuan Huang <laszio@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-by: Gabriel Marin <gmx@chromium.org>