cros-kernel: add config option for Goldfish pipe
Add USE flag "goldfish" that enables support for the Goldfish pipe used
by the Android Emulator for communications between the guest and the
host. In particular, adb uses goldfish to avoid having to specify the IP
address in TCP mode.
When the USE flag is on, enable support for the Goldfish pipe and
keyboard events.

BUG=b:38497992
TEST=emerge-novato chromeos-kernel-4_4

Change-Id: Ic9649b6518a6fe59f26964ac595d97eb0c98eda4
Reviewed-on: https://chromium-review.googlesource.com/525073
Commit-Ready: Nicolas Norvez <norvez@chromium.org>
Tested-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Luis Hector Chavez <lhchavez@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>