Fix reef-uni build: add get_current_board_with_variant non-uni
This adds a paramenter to get_current_board_with_variant that defaults
to false to optionally not drop the -uni suffix off of a board name
while it's returning the variants. get_current_board_with_variant is
used throughout the Portage overlays and scripts so care was taken to
not change the behavior in any existing call site.

BUG=chromium:729604
TEST=emerge-reef-uni --nodeps chromeos-base/dev-install

Change-Id: I8036740a8b399a57c311690694a2a713094a1e65
Reviewed-on: https://chromium-review.googlesource.com/524900
Commit-Ready: Simon Glass <sjg@chromium.org>
Tested-by: Jason Clinton <jclinton@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>