cros-kernel2: Add support for private patches
Under some circumstances it may be necessary to apply a series of
private patches to the kernel. Add support for it.

With this patch, the kernel build system will look for private patches
in directory "${ROOT}/usr/share/chromeos-kernel-patches". The matching
chromeos-kernel-patches-<version> ebuilds are expected to drop
applicable patches in that location.

Patches are only applied if the kernel has been not been checked out
using cros_workon. If the kernel has been checked out, patches will
only be applied if the apply_patches use flag is set.

BUG=chromium:731285
TEST=Build image with patches in private overlays

Change-Id: Ic0f98db893936c9e6868a8a8894465b8acaa6e31
Signed-off-by: Guenter Roeck <groeck@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/528472
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>