net-print/cups: Add readlink to the cupsd seccomp whitelist
Allow readlink for fontconfig which is used used by pdftops.
readlink is used to query /proc/self/exec to obtain the path
of the executable to find system fonts. Without this change
pdftops crashes when attempting to render fonts not embeded
in the PDF.

BUG=chromium:730988
TEST=Print a PDF that requires system fonts on hana and kip.

Change-Id: Ie0338a60d49e1da3d2a8b4be84199f3fec3818e1
Reviewed-on: https://chromium-review.googlesource.com/529985
Commit-Ready: Sean Kau <skau@chromium.org>
Tested-by: Sean Kau <skau@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>