libchromeos USE flags: add arc_migration_policy_check
If arc_migration_policy_check is set, then the device policy
DeviceEcryptfsMigrationStrategy must be checked before providing
the option to user to migrate the data. Only if the device is
consumer owned or the policy allows the migration the user should
be able to start data migration.

BUG=chromium:722371
TEST=none
CQ-DEPEND=CL:533236

Change-Id: Id265b03a41a9f92353dd8c394c0d83684c836f5c
Reviewed-on: https://chromium-review.googlesource.com/533216
Commit-Ready: Igor <igorcov@chromium.org>
Tested-by: Igor <igorcov@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>