sys-boot/coreboot: Disable all interactions with ME cleaner
Upstream coreboot integrated me_cleaner into the build to a) make sure
that there is a valid ME image, b) optionally minimize it.

We only integrate the ME blob at a later stage, so the test in a) fails.
Even though we're no fans of the ME we're also not interested in leaving
the ME hanging in some unknown state, so b) is out, too (in case it ever
becomes the default).

BUG=none
TEST=with this, `emerge-eve coreboot` doesn't fail on CL:533094 anymore
(it broke due to the ME validity test)

Change-Id: I79a859bf55368ab812f2c6ceeb769ec18080579c
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Reviewed-on: https://chromium-review.googlesource.com/535697
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>