dev-db/mariadb: Update EAPI to 5
maridb uses cmake-utils.  The latest upstream cmake-utils.eclass requires
EAPI >= 5.  This ebuild is already conforms to EAPI 5, so just bump the
EAPI.

This package appears to only be used by moblab, and has had pretty
significant chromium specific modification, so an upgrade to the latest
upstream version is beyond the scope of this fix.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium:733477
TEST=emerge-guado_moblab mariadb
 => builds ok

Change-Id: Ib6fa7e455c1b2fce0b1a1555e8a5349c699bbf74
Reviewed-on: https://chromium-review.googlesource.com/536542
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>