dev-libs/libdivsufsort: Re-sync with upstream gentoo
The latest upstream cmake-utils.eclass has a strict requirement for all
inheriting ebuilds to use EAPI >= 5.

The libdivsufsort build uses cmake-utils, but its current ebuild uses
EAPI=3.  Let's take this opportunity to resync with latest upstream
gentoo ebuild.

The upstream gentoo ebuild for libdivsufsort has an equivalent fix as
what we have in chromiumos-overlay for fixing up the lib directory, so
we can drop the local patch.

However, we also maintain a not-upstream change to enable the 64-bit
build of libdivsufsort, which is a dependency of dev-util/bsdiff, so we
can't just move back to a vanilla portage-stable ebuild.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium:733477
TEST=emerge-soraka libdivsufsort
  => built ok
TEST=equery-soraka files libdivsufsort
  => includes:
     /usr/include/divsufsort64.h
     /usr/lib64/libdivsufsort.so.3.0.0
TEST=emerge-soraka bsdiff
  => builds ok

Change-Id: I1cbd454157854565f9c1e64635f0b35db669b400
Reviewed-on: https://chromium-review.googlesource.com/538452
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>