Add dependency of libbrillo (for logging)
In this CL, the dependency is modified because the logging
behaviors are modified. In the original design, the STDOUT
and STRERR are used. However, to make the logs more
accessible, the system log is used. To this end, the libbrillo
dependency is added.

BUG=chromium:712706

TEST=Build package for Guado board and test if the update process is
started correctly. Also, check if the logs of the updater are shown
in the system log.

Change-Id: I19c1d6a771a158ce1e193b97cbe0b2c15f5e4ef3
Reviewed-on: https://chromium-review.googlesource.com/538560
Commit-Ready: Zhongze Hu <frankhu@google.com>
Tested-by: Zhongze Hu <frankhu@google.com>
Reviewed-by: Jiwoong Lee <porce@chromium.org>