cr50: check board ID at startup and set default value if needed
We want to make sure that Cr50 devices have a board ID set to match
the expected default.

This patch adds some code to check if the board ID value has not been
set yet, and set it to default if so.

The new Cr50 binary capable of processing board ID commands is being
published, the new version string is as follows:

  cr50_v1.1.6133-fd788bd tpm2:v0.0.287-1a68fe6 cryptoc:v0.0.8-6283eee

BRANCH=none
BUG=b:62421292
TEST=manual

Change-Id: Iba5abf0086a816cf40af9c3fe1f8f12aa0b58753
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/540839
Reviewed-by: Randall Spangler <rspangler@chromium.org>