media-libs/mesa: Always release 'disp' in ClientWaitSync.
eglClientWaitSync acquires egl display lock and holds it
for all the wait for SYNC_FENCE_KHR.

This patch release the locks before waiting for all the fences
(it was already happening for SYNC_REUSABLE_KHR). In this way
calling eglClientWaitSync will not block other egl calls in
other threads.

This is a temporary solution while we have support for
native fence sync on all the platforms.

BUG=chromium:721463
TEST=emerged and ran WebGLAquarium on cyan.

Change-Id: I9bdc10b76b37957f4536d6c2b6acd076c31a671c
Reviewed-on: https://chromium-review.googlesource.com/541696
Commit-Ready: Kristian H. Kristensen <hoegsberg@chromium.org>
Tested-by: Daniele Castagna <dcastagna@chromium.org>
Reviewed-by: Kristian H. Kristensen <hoegsberg@chromium.org>