Remove unnecessary board name from dtsi files
Originally, the design was to have:
<board>-private-model.dtsi
<board>-model.dtsi

However, the board name isn't set correctly, so the target files end up
with:
-private-model.dtsi
-model.dtsi

This was unnecessary complexity since there will only be one file per
image and that file will contain all of the model details, so to fix the
issue, we're just going to remove the board prepend.

The consumer of this is a compiler in an ebuild that generates the
config.dtb, which already pattern matches on *.dtsi, so that code will
remain unaffected.

BUG=b:62997753
TEST=emerge-reef-uni chromeos-base/chromeos-config-bsp

Change-Id: Ic06794c23af42ffd834cd56fd3454430c8b2a5eb
Reviewed-on: https://chromium-review.googlesource.com/550875
Commit-Ready: C Shapiro <shapiroc@google.com>
Tested-by: C Shapiro <shapiroc@google.com>
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Jason Clinton <jclinton@chromium.org>