Disable copy reloc with pie in gcc.
This feature was getting enabled with binutils 2.27.
However, it broke libbrillo unitests.

This is because of the upstream gcc bug for which the patches were not
backported to 4.9.x branches.
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=65248
This can be removed once patches for the bug are backported to CroS
GCC 4.9.2 branch.

BUG=chromium:711461
TEST=libbrillo tests pass (using gcc as the compiler).

Change-Id: Ia1f2c902252b8bd7a50d409ef6f9ae0c7c772c78
Reviewed-on: https://chromium-review.googlesource.com/552871
Commit-Queue: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Manoj Gupta <manojgupta@chromium.org>
Trybot-Ready: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>