Pass model-specific config file for bluetoothd
For unified builds, we need to pass a different model specific config
file to the bluetoothd daemon when it starts.

This changes the startup methods to invoke a wrapper script that calls
mosys to get the model and if there is a model specific config file,
uses that instead of the default config file.

BUG=b:63111731
TEST=deployed on pyro system and verified bluetoothd startup

Change-Id: Ie31948cc7a3ff069d80a1179819d0546c946146e
Reviewed-on: https://chromium-review.googlesource.com/556150
Commit-Ready: C Shapiro <shapiroc@google.com>
Tested-by: C Shapiro <shapiroc@google.com>
Reviewed-by: Simon Glass <sjg@chromium.org>