Fix authpolicy incremental builds
In an earlier change, protofiles was uprev'ed, but only the r-part of
the protofiles.0.0.3-r1.ebuild was inc'ed. This does not trigger a
rebuild of authpolicy and causes it to break when built incrementally.
This CL includes full instructions how to uprev protofiles and fixes
incremental builds.

BUG=None
TEST=emerge-amd64-generic authpolicy

Change-Id: I5a010d9290cca84bb51d4987ddbef1fabab92626
Reviewed-on: https://chromium-review.googlesource.com/558078
Commit-Ready: Lutz Justen <ljusten@chromium.org>
Tested-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>