Fix depthcharge deps issue causing unibuild fails
A dependency was introduced that caused defconfig and distclean make
targets to depend on libpayload.  For non-unified builds, the default
location worked for this, but unified builds, the location needs to be
explicitly passed (just like all of the other targets).

This should fix the current breakages with the reef-uni-release builder.

BUG=b:63392783
TEST=emerge-reef-uni sys-boot/depthcharge

Change-Id: Ia5cb2379dec83eb10c971e96c9e292816fc9c9f5
Reviewed-on: https://chromium-review.googlesource.com/562650
Commit-Ready: C Shapiro <shapiroc@google.com>
Tested-by: C Shapiro <shapiroc@google.com>
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Jason Clinton <jclinton@chromium.org>