Disable copy reloc with pie in gcc.
This feature was getting enabled with binutils 2.27.
However, it broke libbrillo unitests.

This is because of the upstream gcc bug for which the patches were not
backported to 4.9.x branches.
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=65248
This can be removed once patches for the bug are backported to CroS
GCC 4.9.2 branch.

(Forgot to update 9999 ebuild in CL:552871.)

BUG=chromium:711461
BUG=chromium:738188
TEST=libbrillo tests pass (using gcc as the compiler).

Change-Id: I0b138f0b02316f1359972b513394408e00a8eba7
Reviewed-on: https://chromium-review.googlesource.com/563822
Commit-Ready: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>