trunks: Only install libtrunks_test.a for test USE flag.
This test library was getting installed unconditionally,
seemingly as part of now-outdated Brillo stuff
(http://b/24099462). It only appears to be used by other
packages when the "test" USE flag is set.

Stop installing it by default so a followup change to
trunks.gyp can make us stop building it by default, so
another change to this ebuild can make the trunks package
stop unconditionally depending on gmock.

Also make the attestation, chaps, cryptohome, and
tpm_manager packages, all of which depend on
libtrunks_test.a when built for testing, require that trunks
be built with the "test" USE flag when they themselves are
built with it.

BUG=chromium:740549
TEST=built a dev image for a board with tpm2

Change-Id: I824547721d6fc60d215473fac9f50856165dcc64
Reviewed-on: https://chromium-review.googlesource.com/565960
Commit-Ready: Dan Erat <derat@chromium.org>
Tested-by: Dan Erat <derat@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>