libcxx: remove a symbolic link in the source code.
Android souce tree adds a symbolic link of __cxxabi_config.h
from libcxxabi to libcxx. This causes libcxx install a symbolic
link of __cxxabi_config.h which points to a wrong location.
We do not need that symbolic link and this CL removes that.

BUG=chromium:740232
TEST=sudo emerge cross-x86_64-cros-linux-gnu/libcxx
     /usr/x86_64-cros-linux-gnu/usr/include/c++/v1/__cxxabi_config.h
     is a real file instead of a symbolic link.

Change-Id: I1b0a727c778463a1100f24a9ee19df63b5039190
Reviewed-on: https://chromium-review.googlesource.com/567521
Commit-Ready: Yunlian Jiang <yunlian@chromium.org>
Tested-by: Yunlian Jiang <yunlian@chromium.org>
Reviewed-by: Bailey Forrest <bcf@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>