ntfs3g: Don't make the 'ntfs3g' executable setuid.
cros-disks, currently running as root, launches ntfs3g as the ntfs-3g
user with CAP_SYS_ADMIN + CAP_SYS_SETUID + CAP_SYS_SETGID capabilities,
so 'ntfs3g' doesn't need to be a setuid executable. Even if cros-disks
later runs as a unprivileged user, we can run cros-disks and ntfs3g
with sufficient capabilities for ntfs3g to be able to mount things.

In order for non-root cros-disks to work, the 'ntfs3g' executable needs
to have o+x permissions, or the cros-disks user needs to belong to the
ntfs3g group. If 'ntfs3g' is not setuid, we can make it o+x, and we
don't need to modify groups.

BUG=chromium:726036
TEST=ls -l /build/veyron_minnie/usr/bin/ntfs3g, it's not setuid.
TEST=platform_CrosDisks*
TEST=Plug NTFS-formatted device, see contents on Files app.

Change-Id: If265f689da96022a70809229c474941c90629915
Reviewed-on: https://chromium-review.googlesource.com/568806
Commit-Ready: Jorge Lucangeli Obes <jorgelo@chromium.org>
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>