chromeos-chrome: define use_vaapi when vaapi is used
As part of the plan to remove the libva headers from Chromium's source,
we need to selectively enable compilation of VA-dependent files when
libva is actually used. This patch makes sure the vaapi USE tag is
reflected into the GN_ARGS in such a case.

BUG=chromium:742069
TEST=Ran local emerge on chromeos-chrome, confirmed that use_vaapi was
defined in GN args and that related vaapi files were compiled.

Change-Id: I17ae9edd1b67cd4a38f275706cec35a9a3f820b8
Reviewed-on: https://chromium-review.googlesource.com/571620
Commit-Ready: Alexandre Courbot <acourbot@chromium.org>
Tested-by: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>