profile.bashrc: Update for python 3.6
The python-utils-r1.eclass:python_wrapper_setup installs a stub program named
`python2` when PYTHON_COMPAT only lists python-3 programs.  It's designed to
catch bad packages that still use `python2` even though we said not to.

In Chromium OS though, the compiler sysroot_wrapper.hardened still uses
"python2" in its shebang.  With the shim in place, all calls to the
compiler binaries (clang, gcc, etc) will fail when emerging a python-3
only package with the mysterious message:

   python2 is not supported by python3.3

Conveniently, the function cros_post_pkg_setup_python_eclass_hack
removes the python2 shim such that the sysroot_wrapper works as
expected.

Now that we are updating to python3.6, update this helper function to
clear the new directory.  While we are at it, make it backwards compatible
and future proof with a glob (which only works if the string isn't quoted).

BUG=chromium:736322
TEST=emerge-hana vulkan-loader
TEST=./build_packages --board=soraka
TEST=./build_packages --board=hana

Change-Id: I45178328433f366865a84da7e3df9bd5fcb5bf6f
Reviewed-on: https://chromium-review.googlesource.com/580217
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>