sys-devel/llvm: uprev llvm to 4.0.1 (reland)
We were expanding the ${SYSROOT} variable at build time instead
of at runtime, leading to CQ failures.

When creating the llvm-config wrapper, some boards pointed to the wrong
sysroot. For example, the llvm-config wrapper for reef pointed to:

/build/amd64-generic/usr/lib/llvm/bin/llvm-config

instead of

/build/reef/usr/lib/llvm/bin/llvm-config

Let's escape the ${SYSROOT} variable so it points to the correct place.

BUG=chromium:746968
TEST=emerge-reef llvm mesa

Change-Id: Iba3336a353e0c41e5ad0f8e187937a37dc3aaad8
Reviewed-on: https://chromium-review.googlesource.com/580488
Commit-Ready: Gurchetan Singh <gurchetansingh@chromium.org>
Tested-by: Gurchetan Singh <gurchetansingh@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>