Fix activate date with correct deps and exit cond
This script had multiple issues that would cause silent
failures.  However, since there are no tests for this setup and it only
ever needs to succeed once, the issues were never caught.

The first condition was a dependency to tlsdated, which if the service
hadn't started yet, when the dbus message was sent, it would fail and
cause an early termination of the script.

The second was if there was no value set for ActivateDate in vpd, it
would return non-zero, which would cause early termination from the
script.

Functional tests for this are being added in:
https://chromium-review.googlesource.com/c/583939/

BUG=chromium:748153
TEST=tested locally on pryo

Change-Id: I5126d6b084490d62d9ddaecb53d1c2f0879ca5a7
Reviewed-on: https://chromium-review.googlesource.com/583772
Commit-Ready: C Shapiro <shapiroc@google.com>
Tested-by: C Shapiro <shapiroc@google.com>
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Jason Clinton <jclinton@chromium.org>