sys-devel/llvm: use package.use to disable ncurses
This is an outgrowth of CL:580337. We should use package.use
to disable ncurses when building llvm. Note since we have
two llvm ebuilds in the same directory, this will affect both
of them since they both use the ncurses use flag.  However, as
far as I can tell, ncurses is only used to detect if a terminal
supports escape codes:

https://groups.google.com/forum/#!topic/llvm-dev/MqpXVMy6OrM

So the change shouldn't be that major.

BUG=chromium:734282
TEST=emerge-cyan llvm mesa works

Change-Id: I3653324a16501fed983962d1f836970fe66a0b8c
Reviewed-on: https://chromium-review.googlesource.com/585353
Commit-Ready: Gurchetan Singh <gurchetansingh@chromium.org>
Tested-by: Gurchetan Singh <gurchetansingh@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>