Fixed updater script path for unibuilds
Unibuilds have a concept of supporting per model updater scripts, but
this isn't really viable because the architecture to set per model
environment variables doesn't execute until the updater script is
invoked first.

To solve this, we could either update the outer shellball script to
upack itself, then resolve the model, then decide which script to
execute or we can just remove the per model support and only have one
updater script per build.

This first option sort of duplicates the architecture that's already
natively present in the script itself and in reality, unibuilds are only
actually supported in updater4.sh anyway.

So instead, we're just going to set the script at a per board level for
now and if a compelling use case for per model support comes up, we can
address it then.

Also, in a subsequent change, I'm going to remove the per model script
setting in the dtsi config file in order to remove any implication that
it's supported.

BUG=b:63822433
TEST=locally tested on pyro

Change-Id: I9688a43ccaeb2d795cd8877f75c71bf9d472a88e
Reviewed-on: https://chromium-review.googlesource.com/586888
Commit-Ready: C Shapiro <shapiroc@google.com>
Tested-by: C Shapiro <shapiroc@google.com>
Reviewed-by: Simon Glass <sjg@chromium.org>