upstart: Create dircrypto keyring with owner "cryptohome"
This CL creates the dircrypto keyring with owner user and group as
"cryptohome". This is necessary in order to run cryptohome as non-root
user "cryptohome", as cryptohome needs to create keys in the dircrypto
keyring. The keyring was originally owned (and could only be modified
by) root.

BUG=chromium:741786
TEST=pre-cq
CQ-DEPEND=CL:568778,CL:596670

Change-Id: I0e5152e150fdca1ac4208aae68c4b5c2457003f0
Reviewed-on: https://chromium-review.googlesource.com/590705
Commit-Ready: Matthew Denton <mpdenton@google.com>
Tested-by: Shuqian Zhao <shuqianz@chromium.org>
Tested-by: Matthew Denton <mpdenton@google.com>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>