linux-headers: Export SOCK_DESTROY macro
This is used by the SOCK_DIAG netlink subsystem to kill
connections. We want to be able to use this functionality
from the Chrome OS userspace so we can work around the
cellular IPv6 issues. It first appears in kernel 4.5 so
once linux-headers is uprevved again, this patch will
no longer be necessary.

BUG=chromium:726815
TEST=emerge, check /build/${BOARD}/usr/include

Change-Id: I2b3fc3ab5945b5a6a9f9cbfe1e7700d71363f039
Reviewed-on: https://chromium-review.googlesource.com/592069
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Kevin Cernekee <cernekee@chromium.org>