LLVM: Release version r305632.
This CL updates LLVM in ChromeOS to r305632.
In addition, compiler-rt is enabled as the default compiler
runtime library for llvm.

Links to trybot jobs:
https://docs.google.com/a/google.com/spreadsheets/d/1xVLkraenV8I2EV-UzHArghnBoDUYMX5LiwfxMZ8qmZk/edit?usp=sharing

Autofdo ebuild is updated to ensure that it works both with the current
llvm in paladin builders and the new llvm release.

BUG=chromium:752642
TEST=https://docs.google.com/a/google.com/spreadsheets/d/1xVLkraenV8I2EV-UzHArghnBoDUYMX5LiwfxMZ8qmZk/edit?usp=sharing
TEST=sudo emerge llvm clang works.
TEST=sudo emerge autofdo works with current/old llvm.

Change-Id: I9d6f5dcfe7a81760fc3a2450666fb36b96374257
Reviewed-on: https://chromium-review.googlesource.com/602683
Reviewed-by: Luis Lozano <llozano@chromium.org>
Tested-by: Manoj Gupta <manojgupta@chromium.org>
Trybot-Ready: Manoj Gupta <manojgupta@chromium.org>