Fix app-benchmarks/punybench to build with LLVM.
Punybench uses the "-pg" compilation flag (profiling), which causes
LLVM to add a call to 'mcount' (on arm architectures), which is then
unresolved at link-time.  Adding the flag "--meabi gnu" causes LLVM to
generate calls to __gnu_mcount_nc, instead of to mcount, which the
linker can resolve.  This allows punybench to build with LLVM.

BUG=None
TEST=Successfully emerged punybench for daisy, elm and falco, both with
clang and with forcing the compiler to GCC.

Change-Id: I02f5d0dd1edc4b557d4ef0889e154f995e4cd027
Reviewed-on: https://chromium-review.googlesource.com/604711
Commit-Ready: Caroline Tice <cmtice@chromium.org>
Tested-by: Caroline Tice <cmtice@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>