squashfs-tools: Generate squashfs image file map using unsquashfs
In order to generate a delta update for a squashfs file, we need to be
able to locate the files inside the image along with location of
compressed blocks. Generally squashfs breaks a file into same size
chunks and compresses them. It keeps the location of these compressed
blocks inside its metadata. This CL modifies unsquashfs to generate a
file map when given the flag -m or -map. It does not show the metadata
of the image. The metadata (including the superblock) can be inferred by
subtracting all files locations from the image itself.

The structure of the file map is a list of per-line following:

path start compressed_block_size_1 ... compressed_block_size_N

where:
- 'path' is the full path to the file.
- 'start' is the start byte address of the file.
- 'compressed_block_size_i' is a the length of the i'th compressed block
in bytes. If the block is uncompressed its 25th LSB is set.

It also first prints the list of all fragments along with their address
and block sizes at the beginning for the aforementioned list.

The usage:
unsquashfs -m(ap) <path_to_file_map> <squashfs_image>

BUG=chromium:755201
TEST=unsquashfs -m filemap.txt 25-4k-gzip.img

Change-Id: I4898661bdfc776ac43dd3665f0ca109909000f9d
Reviewed-on: https://chromium-review.googlesource.com/619307
Commit-Ready: Amin Hassani <ahassani@chromium.org>
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>