squashfs-tools: Allow mksquashfs 4K aligns the files in the image
Currently the update engine requires files to be 4k-aligned in order to
create an efficient delta payload. However files inside a squashfs image
are not necessarily 4k (4096) aligned. This patch starts each file in a
4k aligned address and pads zero to the end of the file until it reaches
the next 4k aligned address. This will not change the size of the
compressed blocks (especially the last one) and hence it will not change
how the files are being loaded in kernel or unsquashfs. However on
average this increases the size of the squashfs image which can be
calculated by the following formula:

increased_size = (number_of_unfragmented_files_in_image + number of fragments) * 2048

The 4k alignment can be enabled by flag -4k-align

BUG=chromium:755194
TEST=Mounted a 4k-aligned squashfs image and compared the files with the
original image.

Change-Id: If25eae7af7a8d3207d0d4a1fabf81eec0f69ae03
Reviewed-on: https://chromium-review.googlesource.com/619308
Commit-Ready: Amin Hassani <ahassani@chromium.org>
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>