devserver: migrate update_payload scripts to a new ebuild
Currently update_payload scripts are installed in devserver ebuild. But
since we are moving the use of update_payload scripts to the
update_engine itself, we need a new ebuild (update_payload) to install
these scripts and set it as one of the devserver dependencies.  This
patch increases the devserver version to 3.0.0.

Removes paycheck.py from cros-devutils since it is not used anywhere and
we have a proper copy of it in update_engine repo.

BUG=chromium:737210
TEST=submitted a tryjob; cros flash <ip> <image>
CQ-DEPEND=CL:598435,CL:598434

Change-Id: Ie8c9bb1d2239fc12c1c054b87034a3dfee7596ae
Reviewed-on: https://chromium-review.googlesource.com/614389
Commit-Ready: Amin Hassani <ahassani@chromium.org>
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>