cros-kernel2: Pass most stuff to kernel make as env vars, not cmdline args
In Makefiles, there are two ways to pass variables in from the command
line.  You can either pass them in via environment variables or via
command line arguments.  The two ways of passing in variables causes
different results if the Makefile tries to modify the variables.  For
instance, if you have a Makefile like:

FOODS += and cookies
all:
	@echo "I like $(FOODS)"

And you call it like this:
    FOODS="ice cream" make
...you'll see "I like ice cream and cookies".  ...but if you call:
    make FOODS="ice cream"
...you'll see "I like ice cream".

The fact that the two ways of passing variables behaves so differently
means that any sufficiently complex build system needs to document
when it expects you to use one method and when it expects you to use
the other.  The kernel build system in kbuild.txt [1] explicitly
documents this type of stuff.  Specifically these are listed under
"Environment variables":
* KCFLAGS
* ARCH
* CROSS_COMPILE

...and also the docs talk about "O=" being set via command line.  We
switch all the variables listed as "Environment variables" to be set
via environment variables.

The docs don't explicitly talk about a way to pass in LDFLAGS, perhaps
because the kernel expects to handle all important LDFLAGS itself.
You can pass LDFLAGS explicitly for the final link step of modules
and/or vmlinux, but not general LDFLAGS that are used for creating
temporary object files.  However:

* If you pass LDFLAGS in via command line, it messes up a few use
  cases.  See crbug.com/754462.
* If you pass LDFLAGS in via environment variable, we have the ability
  to pass in some custom LDFLAGS without messing up the use cases.

For now, it seems sane to pass LDFLAGS in via environment variable.
The LDFLAGS that we seem to currently specify don't seem terribly
important and don't seem to affect the build one way or the other, but
it doesn't hurt to pass them in.

[1] https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/Documentation/kbuild/kbuild.txt

BUG=chromium:754462
TEST=Much testing across many machines

Change-Id: I90f21fd6b3f87b893a85a8edf9393d16631c3549
Reviewed-on: https://chromium-review.googlesource.com/624174
Commit-Ready: Douglas Anderson <dianders@chromium.org>
Tested-by: Guenter Roeck <groeck@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>