cros-firmware: Iterate over all models for shared firmware
At present we iterate over only the build targets, so the firmware update
contains only the firmware for the un-shared board. For coral, for
example, this means we get firmware for coral but not astronaut.

Update the eclass to fix this.

BUG=b:64842314
BUG=chromium:741043
TEST=emerge-coral chromeos-config-bsp chromeos-config chromeos-firmware-coral
Copy chroot/build/coral/usr/sbin/chromeos-firmwareupdate to astronaut
./chromeos-firmwareupdate --mode=autoupdate --wp=1 -d -v

Change-Id: Idcbe6db23a515d93ad43af6bd32fecef449057d5
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/627264
Reviewed-by: Jason Clinton <jclinton@chromium.org>
Reviewed-by: C Shapiro <shapiroc@google.com>