openssh-server: use -D in init script
With newer versions of OpenSSH, it has logic[1] to see if it's already
been daemonized, and then changes its forking behavior based on that.
The way Upstart spawns processes triggers that logic which causes the
init script to misbehave -- it can't track the right child process.

Lets change how we start up sshd instead by telling it to always run
in the foreground.  That bypasses the new logic, and simplifies the
overall startup.  Everything sshd was doing to daemonize Upstart is
already doing the same, so there's no need to duplicate.

[1]: https://github.com/openssh/openssh-portable/commit/7fc4766ac78abae81ee75b22b7550720bfa28a33

BUG=chromium:739357
TEST=rebooting/restarting and openssh init still works
TEST=precq passes

Change-Id: I8a477f93bdf6d6f692950ffda9a6bfb956c723a7
Reviewed-on: https://chromium-review.googlesource.com/629048
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>