Reduce dependencies of firmware ebuilds
The ebuilds of the packages used to build firmware images
don't have dependencies of target packages (As firmware runs
when those are out of reach), but depend mostly on packages
being available on the build host (but those are defined in
target-chromium-os-sdk).
Thus, remove the bogus dependencies from the firmware ebuilds.

BUG=none
BRANCH=none
TEST=building firmware still works

Change-Id: I4b03251049c538393c69acf2e3ffab9691be5dd3
Reviewed-on: https://chromium-review.googlesource.com/630637
Commit-Ready: Stefan Reinauer <reinauer@google.com>
Tested-by: Stefan Reinauer <reinauer@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>