coreboot: Display dev mode UI on kahlee.
Kahlee needs a couple of extra CONFIG options set to call into the
option ROM for the graphics device and to set up the framebuffer for
coreboot.

The developer mode image and text are now displayed, and extra
information is provided when the "tab" key is pressed.

BUG=b:64765766,b:63268267
TEST=Manually booted, saw dev mode UI.

Change-Id: I2fb7c593035aefa7d6ff86a804ea3de4c52554d3
Reviewed-on: https://chromium-review.googlesource.com/634212
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Tested-by: Aaron Durbin <adurbin@chromium.org>
Trybot-Ready: Aaron Durbin <adurbin@chromium.org>
Commit-Queue: Justin TerAvest <teravest@chromium.org>