chromeos-bootimage: Drop dependencies of sub packages
chromeos-ec is only a dependency of coreboot. No need to repeat them
here. The ARM dependencies exynos-pre-boot and tegra-bct are obsolete
and no longer needed on ToT. And vboot_reference is a host dependency,
not a target dependency. Remove the dependencies, and unneeded USE
flags.

BUG=none
BRANCH=none
TEST=emerge-fizz chromeos-bootimage succeeds

Change-Id: I7aee590791bea4ba910832caf52261c796d044de
Reviewed-on: https://chromium-review.googlesource.com/647317
Commit-Ready: Stefan Reinauer <reinauer@google.com>
Tested-by: Stefan Reinauer <reinauer@google.com>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>