arc-librm: Put datadir under output prefix.
The new 2.4.83 version will put a new file in a subdirectory of datadir,
but it puts datadir under /usr/share instead of /opt/android/vendor with
the rest of the installed files.  This CL specifically sets datadir to
avoid conflicting with libdrm (non-arc).  This CL is a no-op until the
chromeos-2.4.83 libdrm branch becomes current, but is needed to avoid
build conflicts that would otherwise come up at that point.

BUG=b:65285754
TEST=./build_packages --board=eve --withdev

Change-Id: Ida9ad38849576781926225a6a77de076228cca09
Reviewed-on: https://chromium-review.googlesource.com/651071
Commit-Ready: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>