Add video_WebRtcMainFeedSwitching test to ebuild.
This test starts 5 high resolution loopback peer connections. Four of
the video elements are styled to be small, thumbnail sized and one is in
its native size. This is intended to simulate a video conference with
one main feed and several smaller feeds.

The test then frequently swaps the sourceObject of the main feed with a
randomly chosen smaller feed.

BUG=chromium:760983
TEST=Tested against local DUT

Change-Id: Icdd50c46dc0316c380be3414ccac4a76f8dcd104
Reviewed-on: https://chromium-review.googlesource.com/653158
Commit-Ready: Harpreet Grewal <harpreet@chromium.org>
Tested-by: Kristoffer Erlandsson <kerl@google.com>
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>